Lets figure out the aggregate type when we need it instead
